EP-3251 #125 #117 add DataCube.reduce_dimension
